Safari-611.1.10/LayoutTests/webrtc/libwebrtc/release-while-creating-offer-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Candidate must not have both null sdpMid and sdpMLineIndex
PASS successfullyParsed is true

TEST COMPLETE
This test makes sure that RTCPeerConnection will free itself correctly even if released from the network thread.
